Make user data form reusable
