+++ src/components/Event/EventRegistrationModal.tsx
@@ -3,26 +3,45 @@ import EventRegistrationSignIn from "./EventRegistrationSignIn";
 import Modal from "react-modal";
 import {useNavigate} from "react-router-dom";
 import {useAuth} from "../../providers/Auth/AuthProvider";
-import {useState} from "react";
+import {Dispatch, SetStateAction, useState} from "react";
 import EventRegistrationForm from "./EventRegistrationForm";
+import {UserSchema} from "../OnboardingForm/types";
+import {UserDataForm} from "../UserDataForm";
 
 export function EventRegistrationModal(props:
     {
         isModalOpen: boolean,
-        setIsModalOpen: React.Dispatch<React.SetStateAction <boolean>>
+        setIsModalOpen: Dispatch<SetStateAction <boolean>>
     }) {
     const {currentUser} = useAuth();
+    const [userData, setUserData] = useState<UserSchema>({
+        first_name: "",
+        last_name: "",
+        pronouns: "",
+        ubc_student: "yes",
+        student_id: 12345678,
+        why_pm: "",
+        returning_member: "no"
+    });
     const navigateTo = useNavigate();
 
-    const [step, setStep] = useState(currentUser ? 1 : 0);
+    const handleContinueAsGuest = () => setStep(1);
+    const handleSubmitGuest = async (data: UserSchema) => {
+        setUserData(data);
+        setStep(2);
+        console.log(userData);
+    }
+
+    const [step, setStep] = useState(currentUser ? 2 : 0);
     const stepComponents = [
         <EventRegistrationSignIn
             isOpen={props.isModalOpen}
             onRequestClose={() => props.setIsModalOpen(false)}
             onSignInOrCreateAccount={() => navigateTo("/")}
-            onContinueAsGuest={() => setStep(1)}/>,
+            onContinueAsGuest={handleContinueAsGuest}/>,
+        <UserDataForm onSubmit={handleSubmitGuest} excludeReturningAndWhyPM={true}/>,
         <EventRegistrationForm/>
-    ]
+    ];
 
     return (
         <Modal
