+++ src/components/UserDataForm.tsx
@@ -1,0 +1,172 @@
+import {UserSchema, UserZodObj} from "./OnboardingForm/types";
+import {useForm} from "react-hook-form";
+import {zodResolver} from "@hookform/resolvers/zod";
+import {useEffect} from "react";
+import FormInput from "./FormInput/FormInput";
+
+type UserDataFormProps = {
+    onSubmit: (data: UserSchema) => Promise<void>
+    excludeReturningAndWhyPM?: boolean
+}
+
+export function UserDataForm({onSubmit, excludeReturningAndWhyPM}: UserDataFormProps) {
+    const {
+        register,
+        unregister,
+        handleSubmit,
+        watch,
+        formState: { errors },
+    } = useForm<UserSchema>({
+        defaultValues: {
+            why_pm: "-",
+            returning_member: "no"
+        },
+        resolver: zodResolver(UserZodObj)
+    })
+
+    const student_status = watch("ubc_student")
+    useEffect(() => {
+        if (student_status === "no, other uni") {
+            // Other university student
+            unregister("student_id")
+        } else {
+            // Not a university student
+            unregister("student_id")
+            unregister("year")
+            unregister("faculty")
+            unregister("major")
+        }
+    }, [student_status])
+
+    useEffect(() => {
+        if (excludeReturningAndWhyPM) {
+            unregister("returning_member")
+            unregister("why_pm")
+        } else {
+            console.log("nah")
+        }
+    }, [excludeReturningAndWhyPM])
+
+    return (
+        <form autoComplete="off" className="onboarding-form" onSubmit={handleSubmit(onSubmit)}>
+            <div className="form-content">
+                <div className="form-group">
+                    <FormInput
+                        type="text"
+                        name="first_name"
+                        placeholder="First name"
+                        register={register}
+                        error={errors.first_name}
+                    />
+                    <FormInput
+                        type="text"
+                        name="last_name"
+                        placeholder="Last name"
+                        register={register}
+                        error={errors.last_name}
+                    />
+                    <div className={"form-group-sm"}>
+                        <FormInput
+                            type={"text"}
+                            placeholder={"Pronouns"}
+                            name={"pronouns"}
+                            register={register}
+                            error={errors.pronouns}
+                        />
+                    </div>
+                </div>
+
+                <div className="form-select">
+                    <select required className="select-ubcstudent" {...register("ubc_student",{required: "please select a value"})}>
+                        <option value="" hidden>Are you a UBC student?</option>
+                        <option value={"yes"}>Yes, I'm a UBC student.</option>
+                        <option value={"no, other uni"}>No, I'm from another university.</option>
+                        <option value={"no, other"}>No, I'm not a university student.</option>
+                    </select>
+                    {errors.ubc_student && <span>{errors.ubc_student.message}</span>}
+                </div>
+
+                {student_status === "no, other uni" &&
+                    <FormInput
+                        type="text"
+                        placeholder="University"
+                        name="university"
+                        register={register}
+                        error={errors.university}
+                    />
+                }
+
+                {student_status === "yes" &&
+                    <div className="form-group">
+                        <FormInput
+                            type="text"
+                            placeholder="Student number"
+                            name="student_id"
+                            register={register}
+                            error={errors.student_id}
+                        />
+                    </div>
+
+                }
+
+                {student_status !== "no, other" &&
+                    <div className="form-group">
+                        <div className="form-select form-group-sm">
+                            <select required {...register("year",{required: "please select a value"})}>
+                                <option value="" hidden>Year</option>
+                                <option value={"1"}>1</option>
+                                <option value={"2"}>2</option>
+                                <option value={"3"}>3</option>
+                                <option value={"4"}>4</option>
+                                <option value={"5+"}>5+</option>
+                            </select>
+                            {errors.year && <span>{errors.year.message}</span>}
+                        </div>
+
+                        <FormInput
+                            type="text"
+                            placeholder="Faculty"
+                            name="faculty"
+                            register={register}
+                            error={errors.faculty}
+                        />
+                        <FormInput
+                            type="text"
+                            placeholder="Major"
+                            name="major"
+                            register={register}
+                            error={errors.major}
+                        />
+                    </div>
+                }
+
+                {!excludeReturningAndWhyPM &&
+                    <>
+                    <div className="form-select">
+                        <select
+                            required
+                            {...register("returning_member",
+                                { required: "Please select a value." })}>
+                            <option value="" hidden>Are you a returning member?</option>
+                            <option value="yes">Yes, I'm a returning PMC member.</option>
+                            <option value="no">No, I'm new to PMC.</option>
+                        </select>
+                        {errors.returning_member && <span>{errors.returning_member.message}</span>}
+                    </div>
+
+                    <div className="form-group">
+                        <FormInput
+                            type="text"
+                            placeholder="Why Product Management?"
+                            name="why_pm"
+                            register={register}
+                            error={errors.major}
+                        />
+                    </div>
+                    </>
+                }
+            </div>
+            <button className="submit-button pmc-gradient-background" type="submit">Continue</button>
+        </form>
+    )
+}
