+++ src/components/component-theme.css
@@ -43,10 +43,30 @@
     gap: 1rem;
 }
 
+.form-group {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.form-multi-cols .form-group {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: start;
+    align-items: center;
+    gap: 0.75rem;
+}
+
+.form-group .form-group-sm {
+    width: 16rem;
+}
+
 .form-input {
     width: 100%;
     box-sizing: border-box;
-    border-radius: 999rem;
+    border-radius: 100rem;
     border: none;
     padding: 0.5rem 0.75rem;
 }
@@ -61,12 +81,14 @@
 }
 
 .form-select {
-    box-sizing: border-box;
-    width: 100%;
-    border-radius: 100rem;
-    border: none;
-    padding: 0.5rem 0.75rem;
-    border-right: 0.5rem solid transparent;
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    /*box-sizing: border-box;*/
+    /*width: 100%;*/
+    /*border-radius: 100rem;*/
+    /*border: none;*/
+    /*border-right: 0.5rem solid transparent;*/
 }
 
 .form-select-white {
